added unigram generation + did little refactors
i propose that we pursue n-gram language models as described in
https://web.stanford.edu/~jurafsky/slp3/3.pdf. while usually used for
words in a sentence, it can easily be re-oriented to letters in words.
because all words are to be a fixed length, that should significantly
simplify statistical model that comes out. more than likely it'd be a
5- or 4-gram - depending on how the amount of preceeding letters works
out in context. for tomorrow i'll provide an implementation using a
unigram for now. a unigram just disregards all previous context letters
and works solely with basic relative frequency of the letters.