added initial probability determination using the unigram
currently the training set is exclusively the actual dataset that both
our clone and this code use. i don't think any additional training will
be necessary nor practical. i've provided a basic example on how it
works in main(). put into words though;

determine_most_likely_next_word() takes an array of infos - these are
the keys that define the individual arrays of word indices that we've
had for a bit now - so 'e_def_0' and 'p_con_0' as examples. the
intersection of all these arrays is determined and then ranked via the
unigram i built the generation for in the previous commit.

Mischa: you'll just have to figure out how to extract information from
the clone, convert it into the format understood by this module, and
then you should be golden.