implemented slotgram
it's a bigram of a context and one of the 29 valid chars. not the best
solution, the context is represented by a string with one of the chars
replaced with an '_'. not the path i wanna take, especially when the
basic fivegram comes about. i'll think of something.

the main problem is that different words can generate the same context.
as such, there's some deduplication that needs to be done such as to
allow game datasets other than the training set to be workable with the
slotgram. unfortunately, any method that doesn't just leverage setting a
string as a key in a dict and then playing with the weights from there
doesn't work without an amount of overhead from my own derivations.