cleaned up and sped up slotgram
slotgram generation now takes advantage of the info model to not have to
search the entire word list. it just takes the list where its first
letter is defined, e.g. E is the starting letter of the current context
being processed, let's retrieve the indices for 'e_def_0'.

to avoid string duplication i turned the slotgram from a 2d dictionary
into an index + buffer arrangement. you get the index for the weights of
a context by retrieving it from the slotgram_index, which is a
dictionary where its keys are tuples with the following constellation:
(word_idx, slot_idx). the index is necessary to allow multiple words
resulting in the same context to be able to reference the same context
in the slotgram_weights array. the members of that array are
dictionaries with statistical weights for all 29 valid characters.

the index inherently restricts the model to be functional only for stuff
that's been seen before. but in the sense of conditional probability that's
a fine design decision. it's why i won't be bothering with laplace
smoothing any time soon. what's still to do is to implement the fivegram
and to find a way to dump and read weights into and from files. srcer
takes a couple seconds, even with the optimizations in this patch.