implemented model dumping, dropped fivegram for trigram
no amount of data i shoot into the fivegram - and more notably the
slotgram makes them less shit, so i've opted to reduce the context for
there to be any variance in likelihood with which we can make
estimations. i modified the fivegram into a trigram because their logic
for generation is similar. it works pretty good, especially once you
provide a good corpus.

we can now dump the different models into a text file. for now it's just
outputting to stdout to be redirected into a file; but that'll change in
the next patch. the format is similar to doom wad files, just in
plaintext rather than in binary - easier, more intuitive handling for
umlauts, especially in a python context. there's a header that denotes
the model version and redirects you to a directory for the different
kinds of data it can store. how the data is interpreted is contingent on
the implementation for that lump type e.g. UNIGRAM or TRIGRAM_INDEX.

i'll split worble_hax into two modules, one that only does the
generation of the models, and the other that does the probability
calculations to be converted into the service.