split worble_hax into two modules, switched corpus
worble_hax_gen generates the model to be used by worble_hax_service for
probability estimations. the "model" is comprised of several models,
currently limited to the info_model, unigram and trigram. the fivegram
has been dissolved, the slotgram remains in limbo for now. the corpus
that serves as the basis for both modules can be any lexicon of 5-letter
words. the only condition is that they both use THE SAME ONE. previously
it would have been possible to have each use a different one, but now
that the info_model is baked into the model aswell, they HAVE to use the
same underlying data. to assist in versioning the model format specifies
that every model contain a string to denote its version. you set the
version intended to be used for the service in config.py. worble_hax now
has a file MODEL_MANIFEST, where i'll be writing down different
models that are compatible with the current version of worble_hax.

a new corpus with better conformance to the duden was chosen aswell, its
reproducible generator has been placed in worble_supersanitize. it's
mostly the same, just doesn't abstract away logic as hard. refer to the
latest word list on teams if you can't get your wsl working.